requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
together
pydantic
streamlit
python-dotenv 
groq
nltk
bs4